# cmd/swag/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "API Documentation System Management",
        "termsOfService": "https://github.com/ducvubo",
        "contact": {
            "name": "Vũ Đức Bo",
            "url": "https://github.com/ducvubo",
            "email": "vminhduc8@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "system.management.pg.taphoaictu.id.vn",
    "basePath": "/api/v1",
    "paths": {
...
        "/internal-note": {
            "get": {
                "description": "GetAllInternalNote",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Internal Note"
                ],
                "summary": "GetAllInternalNote",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "pageIndex",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ItnNoteTitle",
                        "name": "ItnNoteTitle",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            },
            "post": {
                "description": "CreateInternalNote",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Internal Note"
                ],
                "summary": "CreateInternalNote",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateInternalNoteDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            },
            "patch": {
                "description": "UpdateInternalNote",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Internal Note"
                ],
                "summary": "UpdateInternalNote",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateInternalNoteDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
        "/internal-note/recycle": {
            "get": {
                "description": "GetAllInternalNoteRecycle",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Internal Note"
                ],
                "summary": "GetAllInternalNoteRecycle",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "pageIndex",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ItnNoteTitle",
                        "name": "ItnNoteTitle",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
        "/internal-note/restore/{id}": {
            "patch": {
                "description": "RestoreInternalNote",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Internal Note"
                ],
                "summary": "RestoreInternalNote",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
        "/internal-note/{id}": {
            "get": {
                "description": "FindInternalNote",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Internal Note"
                ],
                "summary": "FindInternalNote",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            },
            "delete": {
                "description": "DeleteInternalNote",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Internal Note"
                ],
                "summary": "DeleteInternalNote",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CreateInternalNoteDto": {
            "type": "object",
            "required": [
                "itn_note_content",
                "itn_note_title",
                "itn_note_type"
            ],
            "properties": {
                "itn_note_content": {
                    "type": "string"
                },
                "itn_note_title": {
                    "type": "string"
                },
                "itn_note_type": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateInternalNoteDto": {
            "type": "object",
            "required": [
                "itn_note_content",
                "itn_note_id",
                "itn_note_title",
                "itn_note_type"
            ],
            "properties": {
                "itn_note_content": {
                    "type": "string"
                },
                "itn_note_id": {
                    "type": "string"
                },
                "itn_note_title": {
                    "type": "string"
                },
                "itn_note_type": {
                    "type": "string"
                }
            }
        },
...
        "response.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "status code",
                    "type": "integer"
                },
                "data": {
                    "description": "dữ liệu trả về"
                },
                "error": {
                    "description": "dữ liệu lỗi"
                },
                "message": {
                    "description": "thông báo lỗi"
                },
                "statusCode": {
                    "description": "HTTP status code",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AccessTokenAuth": {
            "type": "apiKey",
            "name": "x-at-tk",
            "in": "header"
        },
        "ClientIDAuth": {
            "type": "apiKey",
            "name": "id_user_guest",
            "in": "header"
        },
        "RefreshTokenAuth": {
            "type": "apiKey",
            "name": "x-rf-tk",
            "in": "header"
        }
    },
    "security": [
        {
            "ClientIDAuth": []
        },
        {
            "AccessTokenAuth": []
        },
        {
            "RefreshTokenAuth": []
        }
    ]
}